move append_styles function to dom.ts
